# openapi.yaml
components:
  schemas:
...
    BoxRetrievalResults:
      description: A model encapsulating retrieved research data upload boxes and
        the count thereof.
      properties:
        boxes:
          description: The retrieved research data upload boxes
          items:
            $ref: '#/components/schemas/ResearchDataUploadBox'
          title: Boxes
          type: array
        count:
          description: The total number of unpaginated results
          title: Count
          type: integer
      required:
      - count
      - boxes
      title: BoxRetrievalResults
      type: object
    CreateUploadBoxRequest:
      description: Request model for creating a new research data upload box.
      properties:
        description:
          description: Describes the upload box in more detail
          title: Description
          type: string
        storage_alias:
          description: S3 storage alias to use for uploads
          minLength: 1
          title: Storage Alias
          type: string
        title:
          description: Short meaningful name for the box
          minLength: 1
          title: Title
          type: string
      required:
      - title
      - description
      - storage_alias
      title: CreateUploadBoxRequest
      type: object
...
    GrantAccessRequest:
      description: Request model for granting upload access to a user.
      properties:
        box_id:
          description: ID of the upload box
          format: uuid4
          title: Box Id
          type: string
        iva_id:
          description: ID of the IVA verification
          format: uuid4
          title: Iva Id
          type: string
        user_id:
          description: ID of the user to grant access to
          format: uuid4
          title: User Id
          type: string
        valid_from:
          description: Start date of validity
          format: date-time
          title: Valid From
          type: string
        valid_until:
          description: End date of validity
          format: date-time
          title: Valid Until
          type: string
      required:
      - valid_from
      - valid_until
      - user_id
      - iva_id
      - box_id
      title: GrantAccessRequest
      type: object
    GrantWithBoxInfo:
      description: An UploadGrant with the ResearchDataUploadBox title and description.
      properties:
        box_description:
          description: Describes the upload box in more detail
          title: Box Description
          type: string
        box_id:
          description: ID of the upload box this grant is for
          format: uuid4
          title: Box Id
          type: string
        box_title:
          description: Short meaningful name for the box
          title: Box Title
          type: string
        created:
          description: Date of creation of this grant
          format: date-time
          title: Created
          type: string
        id:
          description: Internal grant ID (same as claim ID)
          format: uuid4
          title: Id
          type: string
        iva_id:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: ID of an IVA associated with this grant
          title: Iva Id
        user_email:
          description: The email address of the user
          format: email
          title: User Email
          type: string
        user_id:
          description: Internal user ID
          format: uuid4
          title: User Id
          type: string
        user_name:
          description: Full name of the user
          title: User Name
          type: string
        user_title:
          anyOf:
          - type: string
          - type: 'null'
          description: Academic title of the user
          title: User Title
        valid_from:
          description: Start date of validity
          format: date-time
          title: Valid From
          type: string
        valid_until:
          description: End date of validity
          format: date-time
          title: Valid Until
          type: string
      required:
      - id
      - user_id
      - box_id
      - created
      - valid_from
      - valid_until
      - user_name
      - user_email
      - box_title
      - box_description
      title: GrantWithBoxInfo
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ResearchDataUploadBox:
      description: A class representing a ResearchDataUploadBox.
      properties:
        changed_by:
          description: ID of the user who performed the latest change
          format: uuid4
          title: Changed By
          type: string
        description:
          description: Describes the upload box in more detail
          title: Description
          type: string
        file_count:
          default: 0
          description: The number of files in the box
          title: File Count
          type: integer
        file_upload_box_id:
          description: The ID of the file upload box.
          format: uuid4
          title: File Upload Box Id
          type: string
        file_upload_box_state:
          description: Current state of the file upload box
          enum:
          - open
          - locked
          - archived
          title: File Upload Box State
          type: string
        file_upload_box_version:
          description: A counter indicating resource version
          title: File Upload Box Version
          type: integer
        id:
          description: Unique identifier for the research data upload box
          format: uuid4
          title: Id
          type: string
        last_changed:
          description: Timestamp of the latest change
          format: date-time
          title: Last Changed
          type: string
        size:
          default: 0
          description: The total size of all files in the box
          title: Size
          type: integer
        state:
          description: Current state of the research data upload box
          enum:
          - open
          - locked
          - archived
          title: State
          type: string
        storage_alias:
          description: S3 storage alias to use for uploads
          title: Storage Alias
          type: string
        title:
          description: Short meaningful name for the box
          title: Title
          type: string
        version:
          description: A counter indicating resource version
          title: Version
          type: integer
      required:
      - version
      - state
      - title
      - description
      - last_changed
      - changed_by
      - file_upload_box_id
      - file_upload_box_version
      - file_upload_box_state
      - storage_alias
      title: ResearchDataUploadBox
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: A service providing a web-accessible management and auth layer over
    upload-path file services.
  title: Upload Orchestration Service
  version: 2.0.0
openapi: 3.1.0
paths:
  /access-grants:
    get:
      description: Endpoint to get the list of all upload access grants. Can be filtered
        by user ID, IVA ID, and box ID. Results are sorted by validity, box ID, user
        ID, IVA ID, and grant ID.
      operationId: get_upload_access_grants_access_grants_get
      parameters:
      - description: The internal ID of the user
        in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: The internal ID of the user
          title: User Id
      - description: The ID of the IVA
        in: query
        name: iva_id
        required: false
        schema:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: The ID of the IVA
          title: Iva Id
      - description: The ID of the upload box
        in: query
        name: box_id
        required: false
        schema:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: The ID of the upload box
          title: Box Id
      - description: Whether the grant is currently valid
        in: query
        name: valid
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Whether the grant is currently valid
          title: Valid
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GrantWithBoxInfo'
                title: Response 200 Get Upload Access Grants Access Grants Get
                type: array
          description: Upload access grants have been fetched.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get upload access grants
      tags:
      - ResearchDataUploadBoxes
    post:
      description: Grant upload access to a user for a single research data upload
        box. Users cannot upload any files until they have been granted access to
        a box.
      operationId: grant_upload_access_access_grants_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantAccessRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Upload access granted successfully.
        '422':
          description: Validation error in request body.
      security:
      - HTTPBearer: []
      summary: Grant upload access
      tags:
      - ResearchDataUploadBoxes
  /access-grants/{grant_id}:
    delete:
      description: Revokes an existing upload access grant.
      operationId: revoke_upload_access_grant_access_grants__grant_id__delete
      parameters:
      - in: path
        name: grant_id
        required: true
        schema:
          format: uuid4
          title: Grant Id
          type: string
      responses:
        '204':
          description: Upload access grant has been revoked.
        '404':
          description: The upload access grant was not found.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Revoke an upload access grant
      tags:
      - ResearchDataUploadBoxes
  /boxes:
    get:
      description: Returns a list of research data upload boxes. Results are sorted
        first by locked status (unlocked followed by locked), then by most recently
        changed, then by box ID. Data Stewards have access to all boxes, while regular
        users may only access boxes to which they have been granted upload access.
      operationId: get_research_data_upload_boxes_boxes_get
      parameters:
      - description: Number of research data upload boxes to skip for pagination
        in: query
        name: skip
        required: false
        schema:
          anyOf:
          - minimum: 0
            type: integer
          - type: 'null'
          description: Number of research data upload boxes to skip for pagination
          title: Skip
      - description: Maximum number of research data upload boxes to return
        in: query
        name: limit
        required: false
        schema:
          anyOf:
          - minimum: 0
            type: integer
          - type: 'null'
          description: Maximum number of research data upload boxes to return
          title: Limit
      - description: Filter by state. None returns all boxes.
        in: query
        name: state
        required: false
        schema:
          anyOf:
          - enum:
            - open
            - locked
            - archived
            type: string
          - type: 'null'
          description: Filter by state. None returns all boxes.
          title: State
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoxRetrievalResults'
          description: Research data upload boxes successfully retrieved.
        '422':
          description: Validation error in query parameters.
      security:
      - HTTPBearer: []
      summary: List upload boxes
      tags:
      - ResearchDataUploadBoxes
    post:
      description: Create a new research data upload box to label and track related
        file uploads for a given user.
      operationId: create_research_data_upload_box_boxes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadBoxRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                format: uuid4
                title: Response 201 Create Research Data Upload Box Boxes Post
                type: string
          description: Upload box created successfully.
        '422':
          description: Validation error in request body.
      security:
      - HTTPBearer: []
      summary: Create upload box
      tags:
      - ResearchDataUploadBoxes
  /boxes/{box_id}:
    get:
      description: Returns the details of an existing research data upload box.
      operationId: get_research_data_upload_box_boxes__box_id__get
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid
          title: Box Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchDataUploadBox'
          description: Upload box details successfully retrieved.
        '404':
          description: Upload box not found or access denied.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get upload box details
      tags:
      - ResearchDataUploadBoxes
    patch:
      description: Update modifiable details for a research data upload box, including
        the description, title, and state. When modifying the state, users are only
        allowed to move the state from OPEN to LOCKED, and all other changes are restricted
        to Data Stewards.
      operationId: update_research_data_upload_box_boxes__box_id__patch
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid
          title: Box Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUploadBoxRequest'
        required: true
      responses:
        '204':
          description: Upload box updated successfully.
        '403':
          description: Access denied.
        '404':
          description: Upload box not found.
        '422':
          description: Validation error in request body.
      security:
      - HTTPBearer: []
      summary: Update upload box
      tags:
      - ResearchDataUploadBoxes
  /boxes/{box_id}/accessions:
    patch:
      description: Update a ResearchDataUploadBox.
      operationId: submit_accession_map_boxes__box_id__accessions_patch
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid
          title: Box Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessionMap'
        required: true
      responses:
        '204':
          description: Accession map successfully submitted.
        '403':
          description: Access denied.
        '404':
          description: Upload box not found.
        '422':
          description: Validation error in request body.
      security:
      - HTTPBearer: []
      summary: Map file IDs to accession numbers for files in the upload box
      tags:
      - ResearchDataUploadBoxes
  /boxes/{box_id}/uploads:
    get:
      description: List the details of all files uploads for a research data upload
        box.
      operationId: list_upload_box_files_boxes__box_id__uploads_get
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid
          title: Box Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FileUploadWithAccession'
                title: Response 200 List Upload Box Files Boxes  Box Id  Uploads Get
                type: array
          description: File upload information successfully retrieved.
        '401':
          description: Access denied.
        '404':
          description: Upload box not found.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List files in upload box
      tags:
      - ResearchDataUploadBoxes
...
  /rpc/archive/{box_id}:
    post:
      description: Archives the specified upload box. The box may no longer be modified,
        and files in the box will be moved to permanent storage. The request body
        must include the complete accession map, which is used for confirmation and
        thus cannot include any updates. If the submitted accession map does not match
        what's in the database, if any files have not been successfully re-encrypted,
        if the upload box is still 'open', or if there are any files which do not
        have an assigned accession number, archival will be denied. Only Data Stewards
        may perform this operation.
      operationId: archive_research_data_upload_box_rpc_archive__box_id__post
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid4
          title: Box Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessionMap'
        required: true
      responses:
        '204':
          description: Upload box archival successfully initiated.
        '403':
          description: Access denied.
        '404':
          description: Upload box not found.
        '409':
          description: Accession map does not match database.
        '422':
          description: Validation error in request body.
      security:
      - HTTPBearer: []
      summary: Archive an upload box
      tags:
      - ResearchDataUploadBoxes
...